% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanPackageExample-package.R
\docType{package}
\name{stanPackageExample-package}
\alias{stanPackageExample-package}
\alias{stanPackageExample}
\title{The 'stanPackageExample' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
